create new SignatureVerifier class
