+++ src/Client/Updater.php
@@ -15,6 +15,7 @@ use Tuf\Metadata\RootMetadata;
 use Tuf\Metadata\SnapshotMetadata;
 use Tuf\Metadata\TimestampMetadata;
 use Tuf\RoleDB;
+use Tuf\SignatureVerifier;
 
 /**
  * Class Updater
@@ -45,25 +46,16 @@ class Updater
     protected $durableStorage;
 
     /**
-     * The role database for the repository.
-     *
-     * @var \Tuf\RoleDB
-     */
-    protected $roleDB;
-
-    /**
-     * The key database for the repository.
-     *
-     * @var \Tuf\KeyDB
-     */
-    protected $keyDB;
-
-    /**
      * The repo file fetcher.
      *
      * @var \Tuf\Client\RepoFileFetcherInterface
      */
     protected $repoFileFetcher;
+
+    /**
+     * @var \Tuf\SignatureVerifier
+     */
+    protected $signatureVerifier;
 
     /**
      * Updater constructor.
@@ -125,8 +117,7 @@ class Updater
     {
         $rootData = RootMetadata::createFromJson($this->durableStorage['root.json']);
 
-        $this->roleDB = RoleDB::createFromRootMetadata($rootData);
-        $this->keyDB = KeyDB::createFromRootMetadata($rootData);
+        $this->signatureVerifier = SignatureVerifier::createFromRootMetadata($rootData);
 
         $this->updateRoot($rootData);
 
@@ -139,7 +130,7 @@ class Updater
         $newTimestampContents = $this->repoFileFetcher->fetchFile('timestamp.json', static::MAXIMUM_DOWNLOAD_BYTES);
         $newTimestampData = TimestampMetadata::createFromJson($newTimestampContents);
         // *TUF-SPEC-v1.0.9 Section 5.2.1
-        $this->checkSignatures($newTimestampData);
+        $this->signatureVerifier->checkSignatures($newTimestampData);
 
         // If the timestamp or snapshot keys were rotating then the timestamp file
         // will not exist.
@@ -172,7 +163,7 @@ class Updater
             throw new MetadataException("Expected snapshot version {$snapshotInfo['version']} does not match actual version " . $newSnapshotData->getVersion());
         }
         // TUF-SPEC-v1.0.9 Section 5.3.2
-        $this->checkSignatures($newSnapshotData);
+        $this->signatureVerifier->checkSignatures($newSnapshotData);
 
         if (isset($this->durableStorage['snapshot.json'])) {
             $currentSnapShotData = SnapshotMetadata::createFromJson($this->durableStorage['snapshot.json']);
@@ -291,89 +282,6 @@ class Updater
         }
     }
 
-    /**
-     * Checks signatures on a verifiable structure.
-     *
-     * @param \Tuf\Metadata\MetadataBase $metaData
-     *     The metadata to check signatures on.
-     *
-     * @return void
-     *
-     * @throws \Tuf\Exception\PotentialAttackException\SignatureThresholdExpception
-     *   Thrown if the signature thresold has not be reached.
-     */
-    protected function checkSignatures(MetadataBase $metaData) : void
-    {
-        $signatures = $metaData->getSignatures();
-
-        $roleInfo = $this->roleDB->getRoleInfo($metaData->getType());
-        $needVerified = $roleInfo['threshold'];
-        $haveVerified = 0;
-
-        $canonicalBytes = JsonNormalizer::asNormalizedJson($metaData->getSigned());
-        foreach ($signatures as $signature) {
-            if ($this->isKeyIdAcceptableForRole($signature['keyid'], $metaData->getType())) {
-                $haveVerified += (int) $this->verifySingleSignature($canonicalBytes, $signature);
-            }
-            // @todo Determine if we should check all signatures and warn for
-            //     bad signatures even this method returns TRUE because the
-            //     threshold has been met.
-            if ($haveVerified >= $needVerified) {
-                break;
-            }
-        }
-
-        if ($haveVerified < $needVerified) {
-            throw new SignatureThresholdExpception("Signature threshold not met on " . $metaData->getType());
-        }
-    }
-
-    /**
-     * Checks whether the given key is authorized for the role.
-     *
-     * @param string $keyId
-     *     The key ID to check.
-     * @param string $roleName
-     *     The role name to check (e.g. 'root', 'snapshot', etc.).
-     *
-     * @return boolean
-     *     TRUE if the key is authorized for the given role, or FALSE
-     *     otherwise.
-     */
-    protected function isKeyIdAcceptableForRole(string $keyId, string $roleName) : bool
-    {
-        $roleKeyIds = $this->roleDB->getRoleKeyIds($roleName);
-        return in_array($keyId, $roleKeyIds);
-    }
-
-    /**
-     * @param string $bytes
-     *     The canonical JSON string of the 'signed' section of the given file.
-     * @param string[] $signatureMeta
-     *     The associative metadata array for the signature. Each signature
-     *     metadata array contains two elements:
-     *     - keyid: The identifier of the key signing the role data.
-     *     - sig: The hex-encoded signature of the canonical form of the
-     *       metadata for the role.
-     *
-     * @return boolean
-     *     TRUE if the signature is valid for the.
-     */
-    protected function verifySingleSignature(string $bytes, array $signatureMeta)
-    {
-        // Get the pubkey from the key database.
-        $keyMeta = $this->keyDB->getKey($signatureMeta['keyid']);
-        $pubkey = $keyMeta['keyval']['public'];
-
-        // Encode the pubkey and signature, and check that the signature is
-        // valid for the given data and pubkey.
-        $pubkeyBytes = hex2bin($pubkey);
-        $sigBytes = hex2bin($signatureMeta['sig']);
-        // @todo Check that the key type in $signatureMeta is ed25519; return
-        //     false if not.
-        return \sodium_crypto_sign_verify_detached($sigBytes, $bytes, $pubkeyBytes);
-    }
-
 
     /**
      * Updates the root metadata if needed.
@@ -405,11 +313,10 @@ class Updater
             }
             $nextRoot = RootMetadata::createFromJson($nextRootContents);
             // *TUF-SPEC-v1.0.9 Section 5.1.3
-            $this->checkSignatures($nextRoot);
-            // Update Role and Key databases to use the new root information.
-            $this->roleDB = RoleDB::createFromRootMetadata($nextRoot);
-            $this->keyDB = KeyDB::createFromRootMetadata($nextRoot);
-            $this->checkSignatures($nextRoot);
+            $this->signatureVerifier->checkSignatures($nextRoot);
+            // Signature verifier to use the new root information.
+            $this->signatureVerifier = SignatureVerifier::createFromRootMetadata($rootData);
+            $this->signatureVerifier->checkSignatures($nextRoot);
             // *TUF-SPEC-v1.0.9 Section 5.1.4
             static::checkRollbackAttack($rootData, $nextRoot, $nextVersion);
             $rootData = $nextRoot;
